+++ README.md
@@ -173,6 +173,17 @@ the contents and the ID from the sql table
 - [x] make buf.yamls
 - [x] maybe use connectrpc
 
+## I need to focus on finishing the project into a working demo duck everything else
+
+- [ ] the latest version of the file will just be the file itself
+- [ ] when the client starts check all files in the target directory
+- [ ] try server connection on client start
+
+### Bonus
+
+- [ ] The fileversions need to be diffs in a order that can than
+      reconstructed on fileID request
+
 ## Risks and Challenges
 
 - Identify any potential risks, challenges, or obstacles that may arise.
