--- example.txt
@@ -1,1 +1,0 @@
-This is some text I will send over connectrpc
