+++ internal/types/file.go
@@ -27,12 +27,3 @@ type FileVersion struct {
 	Contents  string    `bson:"contents"`  // Full contents of the file at this version
 	FileID    string    `bson:"file_id"`   // Unique ID for the file
 }
-
-// TODO: when a file is change it can write a change log and then
-// write to the file to update
-type FileChange struct {
-	Type      string // "add" or "remove"
-	Content   string
-	Position  int    // Line number where change occurred
-	VersionID string `bson:"version_id"` // Unique ID for the file
-}
